+++ src/nvim/strings.h
@@ -2,6 +2,10 @@
 #define NVIM_STRINGS_H
 
 #include <stdbool.h>
+#include <stddef.h>
+
+#include "nvim/types.h"
+#include "nvim/eval_defs.h"
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "strings.h.generated.h"
